requirements.txt
asgiref==3.3.4
Django==3.2.4
django-phonenumber-field==5.2.0
Django-Verify-Email==1.0.7
phonenumbers==8.12.25
pytz==2021.1
sqlparse==0.4.1
